+++ src/main/java/com/heartsave/todaktodak_api/auth/dto/request/EmailOtpCheckRequest.java
@@ -1,6 +1,6 @@
 package com.heartsave.todaktodak_api.auth.dto.request;
 
-import static com.heartsave.todaktodak_api.common.security.constant.ConstraintConstant.Member.EMAIL_MAX_SIZE;
+import static com.heartsave.todaktodak_api.common.constant.ConstraintConstant.Member.EMAIL_MAX_SIZE;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
