+++ src/main/java/com/heartsave/todaktodak_api/auth/dto/request/SignUpRequest.java
@@ -1,6 +1,6 @@
 package com.heartsave.todaktodak_api.auth.dto.request;
 
-import static com.heartsave.todaktodak_api.common.security.constant.ConstraintConstant.Member.*;
+import static com.heartsave.todaktodak_api.common.constant.ConstraintConstant.Member.*;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
