+++ src/test/java/com/heartsave/todaktodak_api/member/controller/MemberControllerTest.java
@@ -6,8 +6,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.heartsave.todaktodak_api.common.WithMockTodakUser;
 import com.heartsave.todaktodak_api.common.exception.errorspec.MemberErrorSpec;
+import com.heartsave.todaktodak_api.common.security.WithMockTodakUser;
 import com.heartsave.todaktodak_api.common.security.domain.TodakUser;
 import com.heartsave.todaktodak_api.config.TestSecurityConfig;
 import com.heartsave.todaktodak_api.member.dto.request.NicknameUpdateRequest;
